+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v1.2.0
 
-- Improve the low_battery-related code, also take the battery-critically-low warning notification into account.
+- Improve the low_battery feature, also take the battery-critically-low warning notification into account.
 
 ## v1.1.0